Move hasher.
