+++ crates/hasher/Cargo.toml
@@ -4,10 +4,4 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-moon_config = { path = "../config"}
-moon_lang_node = { path = "../lang-node" }
-moon_project = { path = "../project"}
-moon_task = { path = "../task"}
-moon_utils = { path = "../utils"}
-serde = { version = "1.0.140", features = ["derive"] }
 sha2 = "0.10.2"
