+++ crates/hasher/src/lib.rs
@@ -1,3 +1,4 @@
 mod hasher;
 
-pub use hasher::TargetHasher;
+pub use hasher::{create_sha256, Hasher};
+pub use sha2::{Digest, Sha256};
